+++ src/pytest_deselect_if/__init__.py
@@ -1,0 +1,3 @@
+from pytest_deselect_if.plugin import deselect_if  # noqa F401
+
+__all__ = ["deselect_if"]