+++ src/pytest_uncollect_if/__init__.py
@@ -1,1 +1,0 @@
-from pytest_uncollect_if.plugin import uncollect_if  # noqa F401