+++ tests/test_deselect_if.py
@@ -3,8 +3,8 @@ from pytest import ExitCode
 
 
 def dont_use_mark_alias(code):
-    newcode = code.replace("from pytest_uncollect_if import uncollect_if", "")
-    newcode = newcode.replace("@uncollect_if", "@pytest.mark.uncollect_if")
+    newcode = code.replace("from pytest_deselect_if import deselect_if", "")
+    newcode = newcode.replace("@deselect_if", "@pytest.mark.deselect_if")
     return newcode
 
 
@@ -15,8 +15,8 @@ mark_alias = pytest.mark.parametrize(
 )
 
 
-def test_run_without_uncollect_marker(pytester):
-    """Test should run normally without the uncollect_if marker."""
+def test_run_without_deselect_marker(pytester):
+    """Test should run normally without the deselect_if marker."""
     pytester.makepyfile(
         """
         def test_simple_case():
@@ -34,14 +34,14 @@ def test_run_without_uncollect_marker(pytester):
 
 @mark_alias
 def test_not_parametrized_class(pytester, transform_code):
-    """Test that an error is raised when the uncollect_if marker is used on a non-parametrized test."""
+    """Test that an error is raised when the deselect_if marker is used on a non-parametrized test."""
     pytester.makepyfile(
         transform_code(
             """
         import pytest
-        from pytest_uncollect_if import uncollect_if
+        from pytest_deselect_if import deselect_if
 
-        @pytest.mark.uncollect_if(func=lambda **kwargs: True)
+        @pytest.mark.deselect_if(func=lambda **kwargs: True)
         class TestTestClass:
             def test_not_parametrized(self):
                 assert 1 == 1
@@ -51,7 +51,7 @@ def test_not_parametrized_class(pytester, transform_code):
     result = pytester.runpytest("-v")
     result.stdout.re_match_lines(
         [
-            r".*ValueError: uncollect_if can only be run on parametrized tests.*",
+            r".*ValueError: deselect_if can only be run on parametrized tests.*",
         ]
     )
     assert result.ret != 0
@@ -59,17 +59,17 @@ def test_not_parametrized_class(pytester, transform_code):
 
 @mark_alias
 def test_class_decoration_applies_to_functions(pytester, transform_code):
-    """Test that the uncollect_if marker can be applied to a class and will apply to all functions in the class."""
+    """Test that the deselect_if marker can be applied to a class and will apply to all functions in the class."""
     pytester.makepyfile(
         transform_code(
             """
         import pytest
-        from pytest_uncollect_if import uncollect_if
+        from pytest_deselect_if import deselect_if
 
         param_decorator = pytest.mark.parametrize('param', [1, 2])
 
         @param_decorator
-        @pytest.mark.uncollect_if(func=lambda param, **kwargs: param == 2)
+        @pytest.mark.deselect_if(func=lambda param, **kwargs: param == 2)
         class TestTestClass:
             def test_one(self, param):
                 assert param != 2
@@ -82,7 +82,7 @@ def test_class_decoration_applies_to_functions(pytester, transform_code):
     result = pytester.runpytest("-v")
     result.stdout.re_match_lines(
         [
-            r"^collecting ... collected 2 items$",
+            r"^collecting ... collected 4 items \/ 2 deselected \/ 2 selected$",
             r".*::TestTestClass::test_one\[1\] PASSED.*",
             r".*::TestTestClass::test_two\[1\] PASSED.*",
         ]
@@ -91,17 +91,17 @@ def test_class_decoration_applies_to_functions(pytester, transform_code):
 
 
 @mark_alias
-def test_class_decoration_with_function_uncollectif(pytester, transform_code):
-    """Test that the uncollect_if marker can be applied to a class and will apply to all functions in the class."""
+def test_class_decoration_with_function_deselectif(pytester, transform_code):
+    """Test that the deselect_if marker can be applied to a class and will apply to all functions in the class."""
     pytester.makepyfile(
         transform_code(
             """
         import pytest
-        from pytest_uncollect_if import uncollect_if
+        from pytest_deselect_if import deselect_if
 
         param_decorator = pytest.mark.parametrize('param', [1, 2])
 
-        @pytest.mark.uncollect_if(func=lambda param, **kwargs: param == 2)
+        @pytest.mark.deselect_if(func=lambda param, **kwargs: param == 2)
         class TestTestClass:
             def test_one(self):
                 assert param != 2
@@ -115,26 +115,26 @@ def test_class_decoration_with_function_uncollectif(pytester, transform_code):
     result = pytester.runpytest("-v")
     result.stdout.re_match_lines(
         [
-            r"^.*ValueError: uncollect_if can only be run on parametrized tests$",
+            r"^.*ValueError: deselect_if can only be run on parametrized tests$",
         ]
     )
     assert result.ret == ExitCode.INTERNAL_ERROR
 
 
 @mark_alias
-def test_class_decoration_with_function_uncollectif_with_raise_when_not_parametrized(
+def test_class_decoration_with_function_deselectif_with_raise_when_not_parametrized(
     pytester, transform_code
 ):
-    """Test that the uncollect_if marker can be applied to a class and will apply to all functions in the class."""
+    """Test that the deselect_if marker can be applied to a class and will apply to all functions in the class."""
     pytester.makepyfile(
         transform_code(
             """
         import pytest
-        from pytest_uncollect_if import uncollect_if
+        from pytest_deselect_if import deselect_if
 
         param_decorator = pytest.mark.parametrize('param', [1, 2])
 
-        @pytest.mark.uncollect_if(func=lambda param, **kwargs: param == 2, raise_when_not_parametrized=False)
+        @pytest.mark.deselect_if(func=lambda param, **kwargs: param == 2, raise_when_not_parametrized=False)
         class TestTestClass:
             def test_one(self):
                 assert 2 == 2
@@ -148,7 +148,7 @@ def test_class_decoration_with_function_uncollectif_with_raise_when_not_parametr
     result = pytester.runpytest("-v")
     result.stdout.re_match_lines(
         [
-            r"^collecting ... collected 2 items$",
+            r"^collecting ... collected 3 items \/ 1 deselected \/ 2 selected$",
             r".*::TestTestClass::test_one PASSED.*",
             r".*::TestTestClass::test_two\[1\] PASSED.*",
         ]
@@ -157,21 +157,19 @@ def test_class_decoration_with_function_uncollectif_with_raise_when_not_parametr
 
 
 @mark_alias
-def test_class_parametrized_decoration_on_function_uncollectif(
-    pytester, transform_code
-):
-    """Test that the uncollect_if marker can be applied to a class and will apply to all functions in the class."""
+def test_class_parametrized_decoration_on_function_deselectif(pytester, transform_code):
+    """Test that the deselect_if marker can be applied to a class and will apply to all functions in the class."""
     pytester.makepyfile(
         transform_code(
             """
         import pytest
-        from pytest_uncollect_if import uncollect_if
+        from pytest_deselect_if import deselect_if
 
         param_decorator = pytest.mark.parametrize('param', [1, 2])
 
         @param_decorator
         class TestTestClass:
-            @pytest.mark.uncollect_if(func=lambda param, **kwargs: param == 2)
+            @pytest.mark.deselect_if(func=lambda param, **kwargs: param == 2)
             def test_one(self, param):
                 assert param < 2
 
@@ -183,7 +181,7 @@ def test_class_parametrized_decoration_on_function_uncollectif(
     result = pytester.runpytest("-v")
     result.stdout.re_match_lines(
         [
-            r"^collecting ... collected 3 items$",
+            r"^collecting ... collected 4 items \/ 1 deselected \/ 3 selected$",
             r".*::TestTestClass::test_one\[1\] PASSED.*",
             r".*::TestTestClass::test_two\[1\] PASSED.*",
             r".*::TestTestClass::test_two\[2\] PASSED.*",
@@ -194,14 +192,14 @@ def test_class_parametrized_decoration_on_function_uncollectif(
 
 @mark_alias
 def test_not_parametrized(pytester, transform_code):
-    """Test that an error is raised when the uncollect_if marker is used on a non-parametrized test."""
+    """Test that an error is raised when the deselect_if marker is used on a non-parametrized test."""
     pytester.makepyfile(
         transform_code(
             """
         import pytest
-        from pytest_uncollect_if import uncollect_if
+        from pytest_deselect_if import deselect_if
 
-        @pytest.mark.uncollect_if(func=lambda **kwargs: True)
+        @pytest.mark.deselect_if(func=lambda **kwargs: True)
         def test_not_parametrized():
             assert 1 == 1
         """
@@ -210,7 +208,7 @@ def test_not_parametrized(pytester, transform_code):
     result = pytester.runpytest("-v")
     result.stdout.re_match_lines(
         [
-            r".*ValueError: uncollect_if can only be run on parametrized tests.*",
+            r".*ValueError: deselect_if can only be run on parametrized tests.*",
         ]
     )
     assert result.ret != 0
@@ -218,13 +216,13 @@ def test_not_parametrized(pytester, transform_code):
 
 @mark_alias
 def test_no_func_arg_error(pytester, transform_code):
-    """Test that an error is raised when the uncollect_if marker has no func argument."""
+    """Test that an error is raised when the deselect_if marker has no func argument."""
     pytester.makepyfile(
         transform_code(
             """
         import pytest
 
-        @pytest.mark.uncollect_if
+        @pytest.mark.deselect_if
         def test_no_func_arg():
             assert 1 == 1
         """
@@ -233,25 +231,25 @@ def test_no_func_arg_error(pytester, transform_code):
     result = pytester.runpytest("-v")
     result.stdout.re_match_lines(
         [
-            r".*ValueError: uncollect_if marker must have a func argument.*",
+            r".*ValueError: deselect_if marker must have a func argument.*",
         ]
     )
     assert result.ret != 0
 
 
 @mark_alias
-def test_uncollect_based_on_condition(pytester, transform_code):
-    """Test that a test is uncollected based on the uncollect_if condition."""
+def test_deselect_based_on_condition(pytester, transform_code):
+    """Test that a test is deselected based on the deselect_if condition."""
     pytester.makepyfile(
         transform_code(
             """
         import pytest
-        from pytest_uncollect_if import uncollect_if
+        from pytest_deselect_if import deselect_if
 
         param_decorator = pytest.mark.parametrize('param', [1, 2])
 
         @param_decorator
-        @uncollect_if(func=lambda param, **kwargs: param == 2)
+        @deselect_if(func=lambda param, **kwargs: param == 2)
         def test_condition_met(param):
             assert param != 2
         """
@@ -260,7 +258,7 @@ def test_uncollect_based_on_condition(pytester, transform_code):
     result = pytester.runpytest("-v")
     result.stdout.re_match_lines(
         [
-            r"^collecting ... collected 1 item$",
+            r"^collecting ... collected 2 items \/ 1 deselected \/ 1 selected$",
             r".*::test_condition_met\[1\] PASSED.*",
         ]
     )
@@ -269,17 +267,17 @@ def test_uncollect_based_on_condition(pytester, transform_code):
 
 @mark_alias
 def test_run_when_condition_not_met(pytester, transform_code):
-    """Ensure test runs when the uncollect_if condition is not met."""
+    """Ensure test runs when the deselect_if condition is not met."""
     pytester.makepyfile(
         transform_code(
             """
         import pytest
-        from pytest_uncollect_if import uncollect_if
+        from pytest_deselect_if import deselect_if
 
         param_decorator = pytest.mark.parametrize('param', [3, 4])
 
         @param_decorator
-        @uncollect_if(func=lambda param, **kwargs: param == 2)
+        @deselect_if(func=lambda param, **kwargs: param == 2)
         def test_condition_not_met(param):
             assert param > 2
         """
@@ -298,19 +296,19 @@ def test_run_when_condition_not_met(pytester, transform_code):
 
 @mark_alias
 def test_multiple_parameters_and_conditions(pytester, transform_code):
-    """Test with multiple parameters and a more complex uncollect condition."""
+    """Test with multiple parameters and a more complex deselect condition."""
     pytester.makepyfile(
         transform_code(
             """
         import pytest
-        from pytest_uncollect_if import uncollect_if
+        from pytest_deselect_if import deselect_if
 
         param1_decorator = pytest.mark.parametrize('param1', [1, 2])
         param2_decorator = pytest.mark.parametrize('param2', [1, 3])
 
         @param2_decorator
         @param1_decorator
-        @uncollect_if(func=lambda param1, param2, **kwargs: param1 == param2)
+        @deselect_if(func=lambda param1, param2, **kwargs: param1 == param2)
         def test_complex_condition(param1, param2):
             assert param1 != param2
         """
@@ -319,7 +317,7 @@ def test_multiple_parameters_and_conditions(pytester, transform_code):
     result = pytester.runpytest("-v")
     result.stdout.re_match_lines(
         [
-            r"^collecting ... collected 3 items$",
+            r"^collecting ... collected 4 items \/ 1 deselected \/ 3 selected$",
             r".*::test_complex_condition\[1-3\] PASSED.*",
             r".*::test_complex_condition\[2-1\] PASSED.*",
             r".*::test_complex_condition\[2-3\] PASSED.*",
@@ -330,7 +328,7 @@ def test_multiple_parameters_and_conditions(pytester, transform_code):
 
 @mark_alias
 def test_deselection_through_markers(pytester, transform_code):
-    """Test with multiple parameters and a more complex uncollect condition."""
+    """Test with multiple parameters and a more complex deselect condition."""
     # register mymarker
     pytester.makeini(
         """
@@ -343,7 +341,7 @@ def test_deselection_through_markers(pytester, transform_code):
         transform_code(
             """
         import pytest
-        from pytest_uncollect_if import uncollect_if
+        from pytest_deselect_if import deselect_if
 
         param1_decorator = pytest.mark.parametrize('param1', [1, 2])
         param2_decorator = pytest.mark.parametrize('param2', [1, 3])
@@ -351,7 +349,7 @@ def test_deselection_through_markers(pytester, transform_code):
         @param2_decorator
         @param1_decorator
         @pytest.mark.mymarker
-        @uncollect_if(func=lambda param1, param2, **kwargs: param1 == param2)
+        @deselect_if(func=lambda param1, param2, **kwargs: param1 == param2)
         def test_complex_condition(param1, param2):
             assert param1 != param2
         """
@@ -360,7 +358,7 @@ def test_deselection_through_markers(pytester, transform_code):
     result = pytester.runpytest("-v", "-m", "not mymarker")
     result.stdout.re_match_lines(
         [
-            r"^collecting ... collected 3 items \/ 3 deselected \/ 0 selected$",
+            r"^collecting ... collected 4 items \/ 4 deselected \/ 0 selected$",
         ]
     )
     assert result.ret == ExitCode.NO_TESTS_COLLECTED
@@ -372,6 +370,6 @@ def test_help_message(pytester):
     )
     result.stdout.fnmatch_lines(
         [
-            "@pytest.mark.uncollect_if(func(**kwargs)): tests marked with uncollect_if will not be collected if func(**kwargs) returns True - like skipif but will not output as a skipped test",
+            "@pytest.mark.deselect_if(func(**kwargs)): tests marked with deselect_if will not be collected if func(**kwargs) returns True - like skipif but output will be deselected rather than skipped",
         ]
     )